closes #i9 - support .html from modules out of the box moved web pack task in it's own file
