+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gulp-workspace",
-  "version": "0.0.11",
+  "version": "0.0.12",
   "description": "PoliteJS Workspace Utility Package for GulpJS",
   "main": "index.js",
   "peerDependencies": {
